Extract converter_factory and hcl_write from utils (#9607) [upstream:528ec5d02ef2a696e4c2887551eff5cf1bd535fe]
