+++ cai2hcl/common/converter.go
@@ -2,7 +2,6 @@ package common
 
 import (
 	"github.com/GoogleCloudPlatform/terraform-google-conversion/v5/caiasset"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/zclconf/go-cty/cty"
 )
 
@@ -11,9 +10,6 @@ type Converter interface {
 	// Convert turns assets into hcl blocks.
 	Convert(asset []*caiasset.Asset) ([]*HCLResourceBlock, error)
 }
-
-// Function initializing a converter from TF resource name and TF resource schema.
-type ConverterFactory = func(name string, schema map[string]*schema.Schema) Converter
 
 // HCLResourceBlock identifies the HCL block's labels and content.
 type HCLResourceBlock struct {
