+++ cai2hcl/common/utils.go
@@ -10,7 +10,6 @@ import (
 	"github.com/hashicorp/hcl/hcl/printer"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/provider"
 	"github.com/zclconf/go-cty/cty"
 	ctyjson "github.com/zclconf/go-cty/cty/json"
 )
@@ -55,18 +54,6 @@ func MapToCtyValWithSchema(m map[string]interface{}, s map[string]*schema.Schema
 	return ret, nil
 }
 
-// Initializes map of converters.
-func CreateConverterMap(converterFactories map[string]ConverterFactory) map[string]Converter {
-	tpgProvider := tpg.Provider()
-
-	result := make(map[string]Converter, len(converterFactories))
-	for name, factory := range converterFactories {
-		result[name] = factory(name, tpgProvider.ResourcesMap[name].Schema)
-	}
-
-	return result
-}
-
 func Convert(assets []*caiasset.Asset, converterNames map[string]string, converterMap map[string]Converter) ([]byte, error) {
 	// Group resources from the same tf resource type for convert.
 	// tf -> cai has 1:N mappings occasionally
@@ -103,53 +90,6 @@ func Convert(assets []*caiasset.Asset, converterNames map[string]string, convert
 	}
 
 	return printer.Format(f.Bytes())
-}
-
-func hclWriteBlock(val cty.Value, body *hclwrite.Body) error {
-	if val.IsNull() {
-		return nil
-	}
-	if !val.Type().IsObjectType() {
-		return fmt.Errorf("expect object type only, but type = %s", val.Type().FriendlyName())
-	}
-	it := val.ElementIterator()
-	for it.Next() {
-		objKey, objVal := it.Element()
-		if objVal.IsNull() {
-			continue
-		}
-		objValType := objVal.Type()
-		switch {
-		case objValType.IsObjectType():
-			newBlock := body.AppendNewBlock(objKey.AsString(), nil)
-			if err := hclWriteBlock(objVal, newBlock.Body()); err != nil {
-				return err
-			}
-		case objValType.IsCollectionType():
-			if objVal.LengthInt() == 0 {
-				continue
-			}
-			// Presumes map should not contain object type.
-			if !objValType.IsMapType() && objValType.ElementType().IsObjectType() {
-				listIterator := objVal.ElementIterator()
-				for listIterator.Next() {
-					_, listVal := listIterator.Element()
-					subBlock := body.AppendNewBlock(objKey.AsString(), nil)
-					if err := hclWriteBlock(listVal, subBlock.Body()); err != nil {
-						return err
-					}
-				}
-				continue
-			}
-			fallthrough
-		default:
-			if objValType.FriendlyName() == "string" && objVal.AsString() == "" {
-				continue
-			}
-			body.SetAttributeValue(objKey.AsString(), objVal)
-		}
-	}
-	return nil
 }
 
 func hashicorpCtyTypeToZclconfCtyType(t hashicorpcty.Type) (cty.Type, error) {
